progress upload
I have first made a simple agent as a function named blind_agent() to get a grasp of what I'm trying to do. I use a built in random method to choose from 4 of the possible moves (up,down,left,right). Now I understand that a function can be used to decide moves, which is basically an agent, I think. Then my next move was to make an "agent" that was not random, so I named it less_dumb_agent(x,y,char). Additionally I tried to implement the idea of Weights used in Neural Networks. Each element in the grid/map has a random numerical value, via a 2D list. These are the weights associated with the move of going there. In this code I think I might have successfully given incentive not to hit an X. This is because every time it senses that it has collided with X I set it so that it deducts value from the weight of that move. As of now the agent does not have any reason to reach the goal, so it simply stays in a place moving side to side. 
Problems left to solve:
* to make it want to go to the goal
* to come up with a way of making the agent not consider the place it was before (because it keeps moving back and forward), without reducing its weight permanently.