edit to previous agent
The agent now does not consider its previous state when assessing what move to make next. This is to prevent back and forth loop which happened pretty much every time. Now it is a bit smarter, however it is not goal oriented at all as of now, and from what I have seen it always ends up in a small continual loop after a dozen moves, by loop I mean it's circling around the same place.